+++ CHANGELOG.md
@@ -14,6 +14,7 @@ If there are not many items, just list them sequentially.
 
 * Fix: UI Clean Up: "Add Tags" should not be tappable (#610)
 * UI Clean Up: Remove pressed state for Cover Image within Bot Profile (#615)
+* Fix: UI Clean Up: Remove "Follow Me" CTA (#612)
 * UI Clean Up: Remove "Add Cover Image" Icon from part of the Bot Create flow (#611)